+++ meson.build
@@ -113,19 +113,22 @@ sources = [
   'src/game/game.c',
   'src/game/gamebuf.c',
   'src/game/gameflow.c',
+  'src/game/gun/gun.c',
+  'src/game/gun/gun_misc.c',
+  'src/game/gun/gun_pistols.c',
+  'src/game/gun/gun_rifle.c',
   'src/game/hair.c',
   'src/game/input.c',
   'src/game/inventry.c',
   'src/game/invfunc.c',
   'src/game/invvars.c',
   'src/game/items.c',
-  'src/game/lara.c',
-  'src/game/lara1gun.c',
-  'src/game/lara2gun.c',
-  'src/game/larafire.c',
-  'src/game/laramisc.c',
-  'src/game/larasurf.c',
-  'src/game/laraswim.c',
+  'src/game/lara/lara.c',
+  'src/game/lara/lara_col.c',
+  'src/game/lara/lara_control.c',
+  'src/game/lara/lara_look.c',
+  'src/game/lara/lara_misc.c',
+  'src/game/lara/lara_state.c',
   'src/game/level.c',
   'src/game/lot.c',
   'src/game/music.c',
