+++ src/game/gun.h
@@ -1,0 +1,3 @@
+#pragma once
+
+#include "game/gun/gun.h"
