+++ src/game/gun/gun_pistols.c
@@ -1,14 +1,11 @@
-#include "game/lara.h"
+#include "game/gun/gun_pistols.h"
 
+#include "game/gun/gun_misc.h"
 #include "game/input.h"
 #include "game/sound.h"
-#include "global/types.h"
 #include "global/vars.h"
 
-#include <stddef.h>
-#include <stdint.h>
-
-void DrawPistols(int32_t weapon_type)
+void Gun_Pistols_Draw(LARA_GUN_TYPE weapon_type)
 {
     int16_t ani = g_Lara.left_arm.frame_number;
     ani++;
@@ -16,10 +13,10 @@ void DrawPistols(int32_t weapon_type)
     if (ani < AF_G_DRAW1 || ani > AF_G_DRAW2_L) {
         ani = AF_G_DRAW1;
     } else if (ani == AF_G_DRAW2) {
-        DrawPistolMeshes(weapon_type);
+        Gun_Pistols_DrawMeshes(weapon_type);
         Sound_Effect(SFX_LARA_DRAW, &g_LaraItem->pos, SPM_NORMAL);
     } else if (ani == AF_G_DRAW2_L) {
-        ReadyPistols();
+        Gun_Pistols_Ready();
         ani = AF_G_AIM;
     }
 
@@ -27,7 +24,7 @@ void DrawPistols(int32_t weapon_type)
     g_Lara.right_arm.frame_number = ani;
 }
 
-void UndrawPistols(int32_t weapon_type)
+void Gun_Pistols_Undraw(LARA_GUN_TYPE weapon_type)
 {
     int16_t anil = g_Lara.left_arm.frame_number;
     if (anil >= AF_G_RECOIL) {
@@ -44,7 +41,7 @@ void UndrawPistols(int32_t weapon_type)
     } else if (anil > AF_G_DRAW1) {
         anil--;
         if (anil == AF_G_DRAW2) {
-            UndrawPistolMeshLeft(weapon_type);
+            Gun_Pistols_UndrawMeshLeft(weapon_type);
         }
     }
     g_Lara.left_arm.frame_number = anil;
@@ -64,7 +61,7 @@ void UndrawPistols(int32_t weapon_type)
     } else if (anir > AF_G_DRAW1) {
         anir--;
         if (anir == AF_G_DRAW2) {
-            UndrawPistolMeshRight(weapon_type);
+            Gun_Pistols_UndrawMeshRight(weapon_type);
         }
     }
     g_Lara.right_arm.frame_number = anir;
@@ -84,7 +81,7 @@ void UndrawPistols(int32_t weapon_type)
     g_Lara.torso_y_rot = (g_Lara.right_arm.y_rot + g_Lara.left_arm.y_rot) / 4;
 }
 
-void ReadyPistols(void)
+void Gun_Pistols_Ready(void)
 {
     g_Lara.gun_status = LGS_READY;
     g_Lara.left_arm.x_rot = 0;
@@ -104,7 +101,7 @@ void ReadyPistols(void)
     g_Lara.left_arm.frame_base = g_Objects[O_PISTOLS].frame_base;
 }
 
-void DrawPistolMeshes(int32_t weapon_type)
+void Gun_Pistols_DrawMeshes(LARA_GUN_TYPE weapon_type)
 {
     int16_t object_num = O_PISTOLS;
     if (weapon_type == LGT_MAGNUMS) {
@@ -123,7 +120,7 @@ void DrawPistolMeshes(int32_t weapon_type)
         g_Meshes[g_Objects[O_LARA].mesh_index + LM_THIGH_R];
 }
 
-void UndrawPistolMeshLeft(int32_t weapon_type)
+void Gun_Pistols_UndrawMeshLeft(LARA_GUN_TYPE weapon_type)
 {
     int16_t object_num = O_PISTOLS;
     if (weapon_type == LGT_MAGNUMS) {
@@ -138,7 +135,7 @@ void UndrawPistolMeshLeft(int32_t weapon_type)
     Sound_Effect(SFX_LARA_HOLSTER, &g_LaraItem->pos, SPM_NORMAL);
 }
 
-void UndrawPistolMeshRight(int32_t weapon_type)
+void Gun_Pistols_UndrawMeshRight(LARA_GUN_TYPE weapon_type)
 {
     int16_t object_num = O_PISTOLS;
     if (weapon_type == LGT_MAGNUMS) {
@@ -153,21 +150,21 @@ void UndrawPistolMeshRight(int32_t weapon_type)
     Sound_Effect(SFX_LARA_HOLSTER, &g_LaraItem->pos, SPM_NORMAL);
 }
 
-void PistolHandler(int32_t weapon_type)
+void Gun_Pistols_Control(LARA_GUN_TYPE weapon_type)
 {
     WEAPON_INFO *winfo = &g_Weapons[weapon_type];
 
     if (g_Input.action) {
-        LaraTargetInfo(winfo);
+        Gun_TargetInfo(winfo);
     } else {
         g_Lara.target = NULL;
     }
     if (!g_Lara.target) {
-        LaraGetNewTarget(winfo);
+        Gun_GetNewTarget(winfo);
     }
 
-    AimWeapon(winfo, &g_Lara.left_arm);
-    AimWeapon(winfo, &g_Lara.right_arm);
+    Gun_AimWeapon(winfo, &g_Lara.left_arm);
+    Gun_AimWeapon(winfo, &g_Lara.right_arm);
 
     if (g_Lara.left_arm.lock && !g_Lara.right_arm.lock) {
         g_Lara.head_x_rot = g_Lara.left_arm.x_rot / 2;
@@ -190,10 +187,10 @@ void PistolHandler(int32_t weapon_type)
             (g_Lara.right_arm.y_rot + g_Lara.left_arm.y_rot) / 4;
     }
 
-    AnimatePistols(weapon_type);
+    Gun_Pistols_Animate(weapon_type);
 }
 
-void AnimatePistols(int32_t weapon_type)
+void Gun_Pistols_Animate(LARA_GUN_TYPE weapon_type)
 {
     PHD_ANGLE angles[2];
     WEAPON_INFO *winfo = &g_Weapons[weapon_type];
@@ -205,7 +202,8 @@ void AnimatePistols(int32_t weapon_type)
         } else if (anir == AF_G_AIM_L && g_Input.action) {
             angles[0] = g_Lara.right_arm.y_rot + g_LaraItem->pos.y_rot;
             angles[1] = g_Lara.right_arm.x_rot;
-            if (FireWeapon(weapon_type, g_Lara.target, g_LaraItem, angles)) {
+            if (Gun_FireWeapon(
+                    weapon_type, g_Lara.target, g_LaraItem, angles)) {
                 g_Lara.right_arm.flash_gun = winfo->flash_time;
                 Sound_Effect(winfo->sample_num, &g_LaraItem->pos, SPM_NORMAL);
             }
@@ -230,7 +228,8 @@ void AnimatePistols(int32_t weapon_type)
         } else if (anil == AF_G_AIM_L && g_Input.action) {
             angles[0] = g_Lara.left_arm.y_rot + g_LaraItem->pos.y_rot;
             angles[1] = g_Lara.left_arm.x_rot;
-            if (FireWeapon(weapon_type, g_Lara.target, g_LaraItem, angles)) {
+            if (Gun_FireWeapon(
+                    weapon_type, g_Lara.target, g_LaraItem, angles)) {
                 g_Lara.left_arm.flash_gun = winfo->flash_time;
                 Sound_Effect(winfo->sample_num, &g_LaraItem->pos, SPM_NORMAL);
             }
