+++ src/game/lara/lara.h
@@ -1,0 +1,16 @@
+#pragma once
+
+// Public Lara routines.
+
+#include "global/types.h"
+
+void Lara_Control(void);
+void Lara_SwapMeshExtra(void);
+void Lara_Animate(ITEM_INFO *item);
+void Lara_AnimateUntil(ITEM_INFO *lara_item, int32_t goal);
+void Lara_UseItem(int16_t object_num);
+void Lara_ControlExtra(int16_t item_num);
+void Lara_InitialiseLoad(int16_t item_num);
+void Lara_Initialise(int32_t level_num);
+void Lara_InitialiseInventory(int32_t level_num);
+void Lara_InitialiseMeshes(int32_t level_num);
