+++ src/game/lara/lara_look.h
@@ -1,0 +1,9 @@
+#pragma once
+
+// Lara lookaround routines.
+
+#include "global/types.h"
+
+void Lara_LookLeftRight(void);
+void Lara_LookUpDown(void);
+void Lara_ResetLook(void);
