+++ src/game/objects/ai/trex.c
@@ -182,7 +182,7 @@ void TRex_LaraDeath(ITEM_INFO *item)
     g_LaraItem->goal_anim_state = AS_SPECIAL;
     g_LaraItem->anim_number = g_Objects[O_LARA_EXTRA].anim_index + 1;
     g_LaraItem->frame_number = g_Anims[g_LaraItem->anim_number].frame_base;
-    LaraSwapMeshExtra();
+    Lara_SwapMeshExtra();
 
     g_LaraItem->hit_points = -1;
     g_Lara.air = -1;
